Moved loading to Utils
